pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "photo-score"
version = "0.1.0"
description = "CLI tool for scoring photo collections using vision models"
readme = "CLAUDE.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "typer>=0.9.0",
    "pillow>=10.0.0",
    "pillow-heif>=0.13.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
photo-score = "photo_score.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
